requirements.txt
aiodns==3.2.0
aiohappyeyeballs==2.6.1
aiohttp==3.10.11
aiosignal==1.3.2
async-timeout==5.0.1
attrs==25.3.0
ccxt==4.4.71
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==44.0.2
frozenlist==1.5.0
idna==3.10
multidict==6.2.0
numpy==2.2.4
pandas==2.2.3
propcache==0.3.1
pycares==4.5.0
pycparser==2.22
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
requests==2.32.3
six==1.17.0
typing_extensions==4.13.0
tzdata==2025.2
urllib3==2.3.0
yarl==1.18.3
pycryptodome=3.22.0
numpy==1.26.4